修復編譯報錯

+++ services/audio_service/client/src/renderer_in_client.cpp
@@ -1657,7 +1657,7 @@ int32_t RendererInClientInner::Write(uint8_t *pcmBuffer, size_t pcmBufferSize, u
     CHECK_AND_RETURN_RET_LOG(converter_->CheckInputValid(bufDesc), ERR_INVALID_PARAM, "Write: Invalid input.");
     converter_->Process(bufDesc);
     uint8_t *buffer;
-    size_t bufferSize;
+    uint32_t bufferSize;
     converter_->GetOutputBufferStream(buffer, bufferSize);
     return Write(buffer, bufferSize);
 }
